chore: move hero text type back into component
