--- src/new-components/heroes/types.ts
@@ -1,7 +1,0 @@
-import { ReactNode } from 'react';
-
-export interface HeroWithText {
-  title: string;
-  kicker?: string;
-  children?: ReactNode;
-}
